Implement KAIOReader() wrapper for endpoint files.
Endpoint files appear like regular files to select() and epoll().
That means select() will return the fd immediately, and epoll()
will completely refuse to operate on them.

This effectively means that you cannot know if read() will block
before calling it. Instead, the Linux Kernel AIO API must be used,
via ctypes wrapper to libaio.

The way KAIOReader works is you give it a fd or a file-like object
that implements fileno(). It then sets up the eventfd and AIO context.
KAIOReader.fileno() returns the fd of the eventfd, which will
become readable when an IO operation submitted by io_submit has
completed. KAIOReader.read() returns the read buffer.

This means that from the outside, KAIOReader behaves pretty much
identical to an endpoint file, except that select() and epoll()
work on it.

There are a few gotchas:

1. KAIOReader.submit() must be called manually the first time.
   io_submit must not be called until after the gadget is bound,
   otherwise it will block forever. Since the KAIOReader is
   created before the binding, constructor does not prime the
   first read.

2. KAIOReader.read() returns a minimum of 32 bytes to workaround
   shortcomings of the protocol parser.

3. io_destroy() must only be called once, or EINVAL is raised.
   KAIOReader.close() will try to prevent this from happening.

4. readinto() etc are not supported, only read().